requirements.txt
discord.py==2.3.2
sympy==1.12
pillow==10.3.0
requests~=2.31.0